+++ backend/src/main/java/Grupo6/VoluntariadoEmergencias/entities/Forms/AbilitiesForm.java
@@ -1,13 +1,12 @@
 package Grupo6.VoluntariadoEmergencias.entities.Forms;
 
-import java.util.ArrayList;
 import java.util.List;
 
 public class AbilitiesForm {
     private String token;
-    private List<Long> ability_ids;
+    private List<Integer> ability_ids;
 
-    public AbilitiesForm(String token, List<Long> ability_ids) {
+    public AbilitiesForm(String token, List<Integer> ability_ids) {
         this.token = token;
         this.ability_ids = ability_ids;
     }
@@ -20,11 +19,11 @@ public class AbilitiesForm {
         this.token = token;
     }
 
-    public List<Long> getAbility_ids() {
+    public List<Integer> getAbility_ids() {
         return ability_ids;
     }
 
-    public void setAbility_ids(List<Long> ability_ids) {
+    public void setAbility_ids(List<Integer> ability_ids) {
         this.ability_ids = ability_ids;
     }
 }
